Results are now appended into existing metadata.json
+ Added changes into other plugins to support this behavior
