+++ metamorph/library/messagehub.py
@@ -109,7 +109,7 @@ import logging.config
 import time
 import os
 
-from metamorph.lib.logging_conf import setup_logging, storing_pretty_json
+from metamorph.lib.support_functions import setup_logging, storing_pretty_json
 from ansible.module_utils.basic import AnsibleModule
