+++ metamorph/plugins/morph_pdc.py
@@ -5,7 +5,7 @@ import logging.config
 
 import requests
 
-from metamorph.lib.logging_conf import setup_logging, storing_pretty_json
+from metamorph.lib.support_functions import setup_logging, storing_pretty_json
 
 
 class PDCApiException(Exception):
